requirements.txt
streamlit
pandas
numpy
sklearn.preprocessing
pymongo
